[Feat] Add Effect Controller class
