+++ ItaCH_Smash_Legends/Assets/Script/Player_Effect/EffectController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 542209f2499852e44bd9b8b851a83e5a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
